azurerm: refactored some shared helpers
