+++ providers/azure/databricks.go
@@ -3,32 +3,14 @@ package azure
 import (
 	"context"
 	"log"
-	"strings"
 
 	"github.com/Azure/azure-sdk-for-go/services/databricks/mgmt/2018-04-01/databricks"
-	"github.com/Azure/go-autorest/autorest"
-	"github.com/hashicorp/go-azure-helpers/authentication"
 
 	"github.com/GoogleCloudPlatform/terraformer/terraformutils"
 )
 
 type DatabricksGenerator struct {
 	AzureService
-}
-
-func (g *DatabricksGenerator) appendResourceAs(resources []terraformutils.Resource, itemID string, itemName string, resourceType string) []terraformutils.Resource {
-	prefix := strings.ReplaceAll(resourceType, resourceType, "dbw")
-	suffix := strings.ReplaceAll(itemName, "-", "_")
-	resourceName := prefix + "_" + suffix
-	res := terraformutils.NewSimpleResource(itemID, resourceName, resourceType, g.ProviderName, []string{})
-	resources = append(resources, res)
-	return resources
-}
-
-func (g *DatabricksGenerator) getArgsProperties() (subscriptionID string, authorizer autorest.Authorizer) {
-	subs := g.Args["config"].(authentication.Config).SubscriptionID
-	auth := g.Args["authorizer"].(autorest.Authorizer)
-	return subs, auth
 }
 
 func (g *DatabricksGenerator) listWorkspaces() ([]databricks.Workspace, error) {
@@ -63,7 +45,7 @@ func (g *DatabricksGenerator) listWorkspaces() ([]databricks.Workspace, error) {
 func (g *DatabricksGenerator) createDatabricksWorkspaces(workspaces []databricks.Workspace) ([]terraformutils.Resource, error) {
 	var resources []terraformutils.Resource
 	for _, item := range workspaces {
-		resources = g.appendResourceAs(resources, *item.ID, *item.Name, "azurerm_databricks_workspace")
+		resources = g.appendResourceAs(resources, *item.ID, *item.Name, "azurerm_databricks_workspace", "dbw")
 	}
 	return resources, nil
 }
